requirements.txt
python-dotenv==1.0.0
requests==2.31.0
openai==1.3.7
twilio==8.12.0 
